backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
fastapi==0.116.1
h11==0.16.0
huggingface_hub==0.23.4
idna==3.10
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.0.0
requests==2.32.4
sniffio==1.3.1
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
torch
transformers
accelerate
openai>=1.0.0
google-generativeai
msal==1.24.0
flask-session==0.5.0
